new v4 API and use Go modules
As moving to modules will require a new import path,
we take the opportunity to make some breaking API changes,
although most of the functionality used by our production
code remains compatible.

A summary of the changes is as follows:

- change the type of ValidationFunc to a type that's used internally,
allowing validation functions to be more efficient in parsing arguments
and checking the value type.
- remove the Valid method and function. For a static validation string,
it's always possible to use a struct value, and for a dynamic validation
string, it's questionable whether walidator should be being used instead
of just writing some Go code.
- remove SetValidationFunc and add AddValidation method. This changes
the semantics of SetValidationFunc so that it can't be used to
change validator names retrospectively (this seems like it's likely to be
a bad idea in general).
- make len validation fail on scalar types. Using "len" on an integer or
floating point value seems like it's wrong - the "length" of an integer is
not its value.
- allow min and max to work on unsigned types
- allow uuid to work on named string types
- make required check that maps and slices are non-nil.
- remove the capability to mutate the global Validator instance (via SetTag and SetValidationFunc). This capability could easily lead to unexpected bevavior when unrelated packages try to do the same thing. You can always create your own Validator instance.
- rename NewValidator to New. `walidator.New` seems less stuttery than `walidator.NewValidator`.
- remove ErrorArray and TextErr; rename ErrorMap to Errors, and include the validator kind in the returned information. This opens the way to including more detailed failure information without compromising the ability of callers to introspect into validation errors. We also make the error return type clearer: an Errors value is _always_ returned.

Internal changes:

- we use a table-driven validator test, making almost all the tests consistent, and also increase test coverage to cover all the code.
- integrate "extensions" into the rest of the code. The code has changed enough from the original that it seems reasonable to call them "builtins" now.
- improve documentation to describe the current functionality.